Don't iterate over all keys to honor 'store' clauses
'store' clauses in templates are relatively rare, so it is wasteful to
iterate over all the returned members just to check that relevant key in
the template wants to store the value in a ref.

Instead, we note which keys want to 'store' while setting up the default
values, and only iterate over them at the end of check().